refactor: Stream data 생성 방식 변경
- String의 extension에서 수행하던 역할을 전문 타입을 만들어 독립
- StreamDataFactory
- StreamData를 만들기 위한 접사들을 깊이 1 단계 낮춤 (Affix 삭제)
- StreamData를 생성하기 위한 타입 메서드를 만들어 Data 인스턴스 생성에 활용하도록 함
