+++ StreamChat/StreamChat/Sources/Helpers/Utils/StreamDataFactory.swift
@@ -1,0 +1,42 @@
+//
+//  StreamDataFactory.swift
+//  StreamChat
+//
+//  Created by Ryan-Son on 2021/08/16.
+//
+
+import Foundation
+
+struct StreamDataFactory {
+
+    enum MessageFormat {
+        case join(username: String)
+        case send(message: String)
+        case leave
+    }
+
+    private enum Prefix {
+        static let join = "USR_NAME::"
+        static let send = "MSG::"
+        static let leave = "LEAVE::"
+    }
+
+    enum Infix {
+        static let receive = "::"
+    }
+
+    private enum Suffix {
+        static let endOfMessage = "::END"
+    }
+
+    static func makeStreamData(_ format: MessageFormat) -> Data? {
+        switch format {
+        case .join(let username):
+            return "\(Prefix.join)\(username)\(Suffix.endOfMessage)".data(using: .utf8)
+        case .send(let message):
+            return "\(Prefix.send)\(message)\(Suffix.endOfMessage)".data(using: .utf8)
+        case .leave:
+            return (Prefix.leave + Suffix.endOfMessage).data(using: .utf8)
+        }
+    }
+}
